% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_methodology.R
\name{mod_methodology_ui}
\alias{mod_methodology_ui}
\alias{mod_methodology_server}
\title{Methodology module}
\usage{
mod_methodology_ui(id)

mod_methodology_server(id)
}
\description{
In-app summary of the development methodology, verification
results, and sources.
}
\keyword{internal}
